Move `hash_array` into hash_utils.rs (#807)
